Explain how to use watchr in the README
Support has been added for watchr, and this is very helpful while
walking the path to enlightenment as it keeps the users focus on
the koans and not on the repeated need to run rake after each edit.

However, only an experienced Rubyist would know how to install and
use watchr (or worse, they might just not notice the watchr config
file). So let's add an optional section to the README explaining
how to use watchr.